requirements.txt
certifi==2021.5.30
chardet==4.0.0
idna==2.10
jproperties==2.1.0
mysql-connector-python==8.0.24
protobuf==3.15.8
requests==2.25.1
six==1.15.0
urllib3==1.26.5
